requirements.txt
django
djangorestframework
psycopg2-binary
Pillow
python-decouple
djangorestframework-simplejwt
celery
django-filter
django-rest-framework-social-oauth2
drf-yasg